requirements.txt
apache-airflow==2.7.2
kafka-python==2.0.2
pyspark==3.5.0